% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{add_levels}
\alias{add_levels}
\title{Add new levels to a factor}
\usage{
add_levels(x, levels_to_add)
}
\arguments{
\item{x}{Factor variable}

\item{levels_to_add}{Character vector of new levels}
}
\value{
A factor with the specified levels added.
}
\description{
Add new levels to a factor
}
