% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{merge_factors}
\alias{merge_factors}
\title{Merge one factor into another, preserving order of levels}
\usage{
merge_factors(x, y)
}
\arguments{
\item{x}{The "dominant" factor, values from `x` will be used if they 
are not NA.}

\item{y}{The factor to be merged into `x`, values from this factor will
be used wherever `x` is NA.}
}
\value{
The merged factor.
}
\description{
Merge one factor into another, preserving order of levels
}
