\name{paste_factors}
\alias{paste_factors}
\title{Combine factors by pasting values}
\usage{
paste_factors(x, y, sep = " ")
}
\arguments{
  \item{x,y}{Factors to be combined.}

  \item{sep}{Character string to separate the levels of x
  and y.}
}
\value{
A new factor with values created by pasting the values of
`x` and `y` together.
}
\description{
Needs work
}

