\name{rename_levels}
\alias{rename_levels}
\title{Rename levels of a factor}
\usage{
rename_levels(x, orig_names, new_names)
}
\arguments{
  \item{x}{Factor variable}

  \item{orig_names}{Original level names}

  \item{new_names}{New level names}
}
\value{
A factor with the specified levels renamed.
}
\description{
Rename levels of a factor
}

