\name{shuffle_within_groups}
\alias{shuffle_within_groups}
\title{Shuffle a vector within each group}
\usage{
shuffle_within_groups(x, g)
}
\arguments{
  \item{x}{The vector to be shuffled.}

  \item{g}{A vector of groups.}
}
\value{
A version of x where elements have been shuffled within
each group.
}
\description{
Shuffle a vector within each group
}

