% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{block_adonis}
\alias{block_adonis}
\title{Adonis for nested experimental designs}
\usage{
block_adonis(formula, data, block_var, nperm = 999)
}
\arguments{
\item{formula}{The model to be tested.}

\item{data}{The data frame within which the model is evaluated.}

\item{block_var}{The name of the variable defining the nested groups.}

\item{nperm}{Number of permutations to run.}
}
\value{
An object describing the fit, similar to the output of `lm`.
}
\description{
Variables within the model, as well as the block-level variable, must be in 
the data frame provided.
}
